优化了用户更新信息的逻辑

+++ src/main/java/org/example/todo_list/repository/jpa/TaskRepository.java
@@ -22,6 +22,6 @@ public interface TaskRepository extends JpaRepository<Task, Long> {
 
     boolean existsByName(String name);
 
-    boolean existsById(Long id);
+    boolean existsById(Integer id);
 
 }
